renamed statements,modified main method

+++ src/main/resources/init_db.sql
@@ -1,7 +1,7 @@
 CREATE DATABASE `manufacturer_db` /*!40100 DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci */ /*!80016 DEFAULT ENCRYPTION='N' */;
 
 CREATE TABLE `drivers` (
-                           `id` INT NOT NULL AUTO_INCREMENT,
+                           `id` BIGINT NOT NULL AUTO_INCREMENT,
                            `name` VARCHAR(45) NOT NULL,
                            `license_number` VARCHAR(45) NOT NULL,
                            `is_deleted` TINYINT NOT NULL DEFAULT '0',
@@ -10,7 +10,7 @@ CREATE TABLE `drivers` (
 
 
 CREATE TABLE `manufacturers` (
-                                 `id` INT NOT NULL AUTO_INCREMENT,
+                                 `id` BIGINT NOT NULL AUTO_INCREMENT,
                                  `name` VARCHAR(45) NOT NULL,
                                  `country` VARCHAR(45) NOT NULL,
                                  `is_deleted` TINYINTNOT NULL DEFAULT '0',
